Исправлены еще некоторые замечания

+++ src/main/resources/liquibase/scripts/pet-bot.sql
@@ -3,13 +3,13 @@
 -- changeset annaLihanova:1
 CREATE TABLE IF NOT EXISTS owner
 (
-    chat_id                INTEGER   NOT NULL PRIMARY KEY,
+    chat_id            BIGINT   NOT NULL PRIMARY KEY,
     name              VARCHAR   NOT NULL
     );
 
 CREATE TABLE IF NOT EXISTS pet
 (
-    id                INTEGER   NOT NULL PRIMARY KEY,
+    id                bigint   NOT NULL PRIMARY KEY,
     name              VARCHAR   NOT NULL,
     age               INTEGER   NOT NULL,
     date_adoption     timestamp NOT NULL
